Separates guest and host typesetters, refactors time functionality
